app/dev-requirements.txt
Flask-CORS==4
pytest>=7.4.3,<8
# pytest-cov>=4.1,<5
dcm-backend>=3.0.0,<4
dcm-common[db]>=3.25.0,<4
